+++ src/components/common/index.ts
@@ -19,3 +19,4 @@ export * from './Button';
 export * from './Policy';
 export * from './IconWithLabel/IconWithLabel';
 export { Dropdown } from './Dropdown/Dropdown';
+export { ResponsiveNav } from './ResponsiveNav/ResponsiveNav';