+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 2. feat: Added mixins to create CSS triangles.
 3. feat: Added prop of policy link text in common component `Form`.
 4. feat: Created common component `Dropdown`.
+5. feat: Created common component `ResponsiveNav`.
 
 ## 0.3.0 (2021-04-06)
 1. feat: Added possibility to use button (common component) as standard link.